resolve error

+++ controleur/delete.php
@@ -12,7 +12,9 @@ header("Content-Type: text/html; charset=utf-8");
 			$nomImage = $_POST['nomImage'];
 			$t = $manager->selectionUnique2('articles',array($nomImage),"code_feraud=$codeFeraud");
 			foreach($t[0] as $k2=>$v2){
-				unlink("../../images_pim/media/$v2");
+				if($k2 == $nomImage){
+					unlink("../../images_pim/media/$v2");
+				}
 			}
 			$tab = array(
 				"$nomImage"=>''
@@ -20,7 +22,5 @@ header("Content-Type: text/html; charset=utf-8");
 			$y =  $manager->modifier('articles',$tab,"code_feraud=$codeFeraud");
 			echo json_encode('Suppression effectuée');
 		}
-
-		
 	}
 ?>
